% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{data-cowbw}
\alias{data-cowbw}
\alias{cowbw}
\alias{cowbw2}
\title{Cow body weight}
\format{\describe{
\item{cowid:}{Identifier of cow}
\item{tfc:}{Time from calving in days}
\item{bw:}{Body weight}
}}
\usage{
cowbw
}
\description{
Body weight of cows througout a lactation for 48 cows.
}
\details{
There are 48 cowids in \code{cowbw}; \code{cowbw2}
    contains data for the first six cowids.
}
\examples{

require(ggplot2)
data(cowbw2)

qplot(tfc, bw, data=cowbw2, colour=cowid) + geom_path()

cowlst <- cowbw2 \%>\% split(.$cowid)

i <- 2
dd <- cowlst[[i]]
tvar <- dd$tfc
yvar <- dd$bw

f1 <- ses(yvar, tvar)
f2 <- des(yvar, tvar)

plot(f1)
at <- seq(10, 310, by=20)
forecast_lines(f1, at=at, ahead=0:20, col='red', lwd=3)
forecast_lines(f2, at=at, ahead=0:20, col='blue', lwd=3)
}
\keyword{datasets}
