\name{dataIrony-package}
\alias{dataIrony-package}
%%\alias{dataIrony}
\docType{package}
\title{
\packageTitle{dataIrony}
}
\description{
\packageDescription{dataIrony}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{dataIrony}
\packageIndices{dataIrony}
%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
}
\author{
\packageAuthor{dataIrony}

Maintainer: \packageMaintainer{dataIrony}
}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
