\name{esmooth-package}
\alias{esmooth-package}
%%\alias{esmooth}
\docType{package}
\title{
\packageTitle{esmooth}
}
\description{
\packageDescription{esmooth}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{esmooth}
\packageIndices{esmooth}
%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
}
\author{
\packageAuthor{esmooth}

Maintainer: \packageMaintainer{esmooth}
}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
