% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EsmoothFUN.R
\name{esmooth}
\alias{esmooth}
\alias{ses}
\alias{des}
\title{Exponential smoothing}
\usage{
ses(yvar, tvar = NULL, alpha = NULL, init = NULL, filter. = NULL)

des(yvar, tvar = NULL, alpha = NULL, init = NULL, filter. = NULL)
}
\arguments{
\item{yvar}{Time series, a numeric vector.}

\item{tvar}{Time variable, a numeric vector. Defaults to 1, 2, ...}

\item{alpha}{Smoothing parameter; if NULL, alpha will be estimated
by minimizing a prediction error.}

\item{init}{Intial value for the first smoothed value; defaults to
the average of the first three observations.}

\item{filter.}{Experimental feature, please do not use.}
}
\description{
Exponential smoothing of (possibly non-equidistant)
    time series.
}
\examples{

yvar <- aggregate(co2)
tvar <- seq_along(yvar)
f1 <- ses(yvar, tvar)
f2 <- des(yvar, tvar)

at <- 1 + seq(0, 35, by=5)

par(mfrow=c(2,1))
plot(f1)
forecast_lines(f1, at=at, ahead=0:5, col='red', lwd=3)
forecast_lines(f2, at=at, ahead=0:5, col='blue', lwd=3)

## Add more noise
yvar2 <- yvar + rnorm(length(yvar), sd=40)
f1 <- ses(yvar2, tvar)
f2 <- des(yvar2, tvar)

plot(f1)
at <- 1 + seq(0, 35, by=5)
forecast_lines(f1, at=at, ahead=0:5, col='red', lwd=3)
forecast_lines(f2, at=at, ahead=0:5, col='blue', lwd=3)

}
