% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataIrony-ES.R
\name{forecast}
\alias{forecast}
\alias{forecast.SES}
\alias{forecast.DES}
\alias{forecast_lines}
\alias{forecast_lines.ES}
\title{Forecast for exponential smoothing}
\usage{
forecast(object, at = NULL, h = NULL)

\method{forecast}{SES}(object, at = NULL, h = NULL)

\method{forecast}{DES}(object, at = NULL, h = NULL)

forecast_lines(object, at = NULL, h = NULL, ...)

\method{forecast_lines}{ES}(object, at = NULL, h = NULL, ...)
}
\arguments{
\item{object}{A 'SES' or 'DES' object.}

\item{at}{Timepoint at which prediction starts.}

\item{h}{Steps ahead for which predictions are made.}

\item{...}{Graphics parameters passed on to lines()}
}
\description{
Forecast for exponential smoothing.
}
